+++ src/tmtccmd/tc/packer.py
@@ -10,7 +10,7 @@ from typing import Union
 from tmtccmd.tc.definitions import TcQueueT
 from spacepackets.ecss.tc import PusTelecommand
 from tmtccmd.utility.logger import get_console_logger
-from tmtccmd.tc.service_17_test import pack_service17_ping_command
+from tmtccmd.pus.service_17_test import pack_service_17_ping_command
 from tmtccmd.tc.service_5_event import pack_generic_service5_test_into
 
 LOGGER = get_console_logger()
@@ -39,7 +39,7 @@ class ServiceQueuePacker:
 
 
 def default_single_packet_preparation() -> PusTelecommand:
-    return pack_service17_ping_command(ssc=1700)
+    return pack_service_17_ping_command(ssc=1700)
 
 
 def default_service_queue_preparation(service: Union[str, int], op_code: str, service_queue: TcQueueT):
@@ -48,6 +48,6 @@ def default_service_queue_preparation(service: Union[str, int], op_code: str, se
         pack_generic_service5_test_into(service_queue)
     if service == CoreServiceList.SERVICE_17.value:
         service_queue.appendleft((QueueCommands.PRINT, "Sending ping command PUS TC[17,1]"))
-        service_queue.appendleft(pack_service17_ping_command(ssc=1700).pack_command_tuple())
+        service_queue.appendleft(pack_service_17_ping_command(ssc=1700).pack_command_tuple())
     else:
         LOGGER.warning("Invalid Service !")
